+++ app/src/main/java/com/symeonchen/wakeupscreen/services/reminder/ReminderEngine.kt
@@ -95,7 +95,7 @@ object ReminderEngine {
             return
         }
 
-        ScreenWakeUtils.wakeUpScreen(powerManager)
+        ScreenWakeUtils.wakeUpScreen(appContext, powerManager)
 
         val round = DataInjection.repeatReminderRoundCount + 1
         DataInjection.repeatReminderRoundCount = round